Fixed problem with pointselectorwidget.
User was able to click the widget when no image was shown.  To hide
this, the image is check for null in mousePressEvent.  Still get warning
because the program tries drawing a null image when resizing but to my
knowledge it just throws that error out that can be ignored.